code/requirements.txt
torch==2.0.1
torchvision==0.15.2
openai-clip==1.0.1
pillow==10.0.1
boto3==1.28.60
s3transfer==0.7.0
Unidecode==1.3.7
easyocr==1.7.1
pypdf==4.0.1